requirements.txt
discord.py>=2.4.0
requests>=2.32.0
python-dotenv>=1.0.1
pydantic>=2.7.0
APScheduler>=3.10.4
beautifulsoup4>=4.12.0
